feat: [#55850] Rework `OrgTreeDiff` to produce expected diff
Now that we have mocked versions of both the MO and the SD trees, rework
the `OrgTreeDiff` logic to produce the expected diff output.

This commit changes `RemoveOperation`, `AddOperation` and
`UpdateOperation` from dataclasses to classes implementing an
`Operation` ABC.

This commit changes `OrgTreeDiff` so it produces the `Operation`
subclasses described above.

This commit changes the inner `OrgTreeDiff` logic to be based on two
`DeepDiff` instances: an "ID-based" deep diff using the org unit node
UUIDs, and a "structural" deep diff. This allows `OrgTreeDiff` to
produce the expected diff for the "remove", "update" and "add"
operations.