feat: [#56946] Add `TreeDiffExecutor` handling mutation of MO tree
The `TreeDiffExecutor` takes a GraphQL session and an `OrgTreeDiff`
instance as input. `TreeDiffExecutor` processes the `OrgTreeDiff`
instance, and produces the relevant GraphQL mutations for removing,
updating and adding org units in the MO tree.

`TreeDiffExecutor` can return a list of mutations to execute, or execute
the mutations immediately using its `execute` method.

The relevant mutations are implemented as subclasses of the new
`Mutation` ABC:
- `RemoveOrgUnitMutation` - matches the `RemoveOperation` operation
- `UpdateOrgUnitMutation` - matches the `UpdateOperation` operation
- `AddOrgUnitMutation` - matches the `AddOperation` operation