feat: [#56946] Use SD `ActivationDate`/`DeactivationDate` for validity
This extends the following classes with a new `validity` member:
* `OrgUnit` (representing a MO org unit instance),
* `OrgUnitNode` (representing a "generic" org unit, coming from either
MO or SD),
* `UpdateOperation` (representing the SD validity which should be
written to the corresponding MO org unit),
* `AddOperation` (representing the SD validity of an org unit which
should be created in MO)

Root org units in both the MO and SD trees are assumed to have no
validity.

This adds `types-python-dateutil` as a new dev dependency, and adds it
to the `mypy` hook in `.pre-commit-config.yaml`.