test: [#57506] Adding and updating org units returns the org unit UUID
This brings the test fixture `_MockGraphQLSession` in line with what
actually happens when executing the mutations against a live MO, e.g.
returning the org unit UUID of the added or updated org unit.

This commit also removes the fixture `_MockGraphQLSessionForMutation`,
as testing the `TreeDiffExecutor` requires a fixtures which "responds"
to *both* the GQL queries *and* the mutations.