feat!: remove dry-run from person sync
BREAKING CHANGE:

This commit removes the dry-run feature from person sync, since
it is never used in practice and since it is not working when called
via /minisync/sync-person-and-employment. The reason for the latter
is that the employment dry-run sync requires that the corresponding
person has actually been sync first.